requirements.txt
pandas>=1.3.0
python-docx>=0.8.11
openpyxl>=3.0.0
xlrd>=2.0.0
pyinstaller>=5.0.0
pandas==2.3.0
python-docx==1.2.0
openpyxl==3.1.5
xlrd==2.0.1
pillow==11.2.1
lxml==6.0.0 